pass3 array
claude

+++ shedskin/lib/array.cpp
@@ -45,8 +45,8 @@ template<> template<> void *array<int>::extend(list<__ss_int> *l) {
         case 'I': for(size_t i=0; i<len; i++) *((unsigned int *)(&this->units[pos+i*itemsize])) = (unsigned int)l->units[i]; break;
         case 'l': for(size_t i=0; i<len; i++) *((signed long *)(&this->units[pos+i*itemsize])) = (signed long)l->units[i]; break;
         case 'L': for(size_t i=0; i<len; i++) *((unsigned long *)(&this->units[pos+i*itemsize])) = (unsigned long)l->units[i]; break;
-        case 'q': for(size_t i=0; i<len; i++) *((signed long *)(&this->units[pos+i*itemsize])) = (signed long long)l->units[i]; break;
-        case 'Q': for(size_t i=0; i<len; i++) *((unsigned long *)(&this->units[pos+i*itemsize])) = (unsigned long long)l->units[i]; break;
+        case 'q': for(size_t i=0; i<len; i++) *((signed long long *)(&this->units[pos+i*itemsize])) = (signed long long)l->units[i]; break;
+        case 'Q': for(size_t i=0; i<len; i++) *((unsigned long long *)(&this->units[pos+i*itemsize])) = (unsigned long long)l->units[i]; break;
         case 'f': for(size_t i=0; i<len; i++) *((float *)(&this->units[pos+i*itemsize])) = (float)l->units[i]; break;
         case 'd': for(size_t i=0; i<len; i++) *((double *)(&this->units[pos+i*itemsize])) = (double)l->units[i]; break;
     }
